+++ packages/messenger/src/Messenger.ts
@@ -1,5 +1,3 @@
-import { RestrictedMessenger } from './RestrictedMessenger';
-
 export type ActionHandler<
   Action extends ActionConstraint,
   ActionType = Action['type'],
@@ -64,13 +62,37 @@ export type SelectorEventHandler<SelectorReturnValue = unknown> = (
 ) => void;
 
 export type ActionConstraint = {
-  type: string;
+  type: NamespacedName;
   handler: ((...args: never) => unknown) | ((...args: never[]) => unknown);
 };
 export type EventConstraint = {
-  type: string;
+  type: NamespacedName;
   payload: unknown[];
 };
+
+/**
+ * Extract action types from a Messenger type.
+ *
+ * @template Subject - The messenger type to extract from.
+ */
+type MessengerActions<
+  Subject extends Messenger<string, ActionConstraint, EventConstraint>,
+> =
+  Subject extends Messenger<string, infer Action, EventConstraint>
+    ? Action
+    : never;
+
+/**
+ * Extract event types from a Messenger type.
+ *
+ * @template Subject - The messenger type to extract from.
+ */
+type MessengerEvents<
+  Subject extends Messenger<string, ActionConstraint, EventConstraint>,
+> =
+  Subject extends Messenger<string, ActionConstraint, infer Event>
+    ? Event
+    : never;
 
 /**
  * Metadata for a single event subscription.
@@ -78,6 +100,11 @@ export type EventConstraint = {
  * @template Event - The event this subscription is for.
  */
 type SubscriptionMetadata<Event extends EventConstraint> = {
+  /**
+   * Whether this subscription is for a delegated messenger. Delegation subscriptions are ignored
+   * when clearing subscriptions.
+   */
+  delegation: boolean;
   /**
    * The optional selector function for this subscription.
    */
@@ -117,35 +144,80 @@ export type NotNamespacedBy<
 export type NamespacedName<Namespace extends string = string> =
   `${Namespace}:${string}`;
 
-type NarrowToNamespace<Name, Namespace extends string> = Name extends {
-  type: `${Namespace}:${string}`;
-}
-  ? Name
-  : never;
+/**
+ * A messenger that actions and/or events can be delegated to.
+ *
+ * This is a minimal type interface to avoid complex incompatibilities resulting from generics over
+ * invariant types.
+ */
+type DelegatedMessenger = Pick<
+  // The type is broadened to all actions/events because some messenger methods are contravariant
+  // over this type (`registerDelegatedActionHandler` and `publishDelegated` for example). If this
+  // type is narrowed to just the delegated actions/events, the types for event payload and action
+  // parameters would not be wide enough.
+  Messenger<string, ActionConstraint, EventConstraint>,
+  | '_internalPublishDelegated'
+  | '_internalRegisterDelegatedActionHandler'
+  | '_internalRegisterDelegatedInitialEventPayload'
+  | '_internalUnregisterDelegatedActionHandler'
+>;
 
-type NarrowToAllowed<Name, Allowed extends string> = Name extends {
-  type: Allowed;
-}
-  ? Name
-  : never;
+type StripNamespace<Namespaced extends NamespacedName> =
+  Namespaced extends `${string}:${infer Name}` ? Name : never;
 
 /**
  * A message broker for "actions" and "events".
  *
  * The messenger allows registering functions as 'actions' that can be called elsewhere,
  * and it allows publishing and subscribing to events. Both actions and events are identified by
- * unique strings.
+ * unique strings prefixed by a namespace (which is delimited by a colon, e.g.
+ * `Namespace:actionName`).
  *
  * @template Action - A type union of all Action types.
  * @template Event - A type union of all Event types.
+ * @template Namespace - The namespace for the messenger.
  */
 export class Messenger<
+  Namespace extends string,
   Action extends ActionConstraint,
   Event extends EventConstraint,
+  Parent extends Messenger<
+    string,
+    ActionConstraint,
+    EventConstraint,
+    // Use `any` to avoid preventing a parent from having a parent. `any` is harmless in a type
+    // constraint anyway, it's the one totally safe place to use it.
+    // eslint-disable-next-line @typescript-eslint/no-explicit-any
+    any
+  > = never,
 > {
-  readonly #actions = new Map<Action['type'], unknown>();
+  readonly #namespace: Namespace;
+
+  /**
+   * The parent messenger. All actions/events under this namespace are automatically delegated to
+   * the parent messenger.
+   */
+  readonly #parent?: DelegatedMessenger;
+
+  readonly #actions = new Map<Action['type'], Action['handler']>();
 
   readonly #events = new Map<Event['type'], EventSubscriptionMap<Event>>();
+
+  /**
+   * The set of messengers we've delegated events to and their event handlers, by event type.
+   */
+  readonly #subscriptionDelegationTargets = new Map<
+    Event['type'],
+    Map<DelegatedMessenger, ExtractEventHandler<Event, Event['type']>>
+  >();
+
+  /**
+   * The set of messengers we've delegated actions to, by action type.
+   */
+  readonly #actionDelegationTargets = new Map<
+    Action['type'],
+    Set<DelegatedMessenger>
+  >();
 
   /**
    * A map of functions for getting the initial event payload.
@@ -166,19 +238,65 @@ export class Messenger<
   >();
 
   /**
+   * Construct a messenger.
+   *
+   * If a parent messenger is given, all actions and events under this messenger's namespace will
+   * be delegated to the parent automatically.
+   *
+   * @param args - Constructor arguments
+   * @param args.namespace - The messenger namespace.
+   * @param args.parent - The parent messenger.
+   */
+  constructor({
+    namespace,
+    parent,
+  }: {
+    namespace: Namespace;
+    parent?: Action['type'] extends MessengerActions<Parent>['type']
+      ? Event['type'] extends MessengerEvents<Parent>['type']
+        ? Parent
+        : never
+      : never;
+  }) {
+    this.#namespace = namespace;
+    this.#parent = parent;
+  }
+
+  /**
    * Register an action handler.
    *
    * This will make the registered function available to call via the `call` method.
+   *
+   * The action being registered must be under the same namespace as the messenger.
    *
    * @param actionType - The action type. This is a unique identifier for this action.
    * @param handler - The action handler. This function gets called when the `call` method is
    * invoked with the given action type.
    * @throws Will throw when a handler has been registered for this action type already.
-   * @template ActionType - A type union of Action type strings.
+   * @template ActionType - A type union of Action type strings under this messenger's namespace.
    */
-  registerActionHandler<ActionType extends Action['type']>(
+  registerActionHandler<
+    ActionType extends Action['type'] & NamespacedName<Namespace>,
+  >(actionType: ActionType, handler: ActionHandler<Action, ActionType>) {
+    /* istanbul ignore if */ // Branch unreachable with valid types
+    if (!this.#isInCurrentNamespace(actionType)) {
+      throw new Error(
+        `Only allowed registering action handlers prefixed by '${
+          this.#namespace
+        }:'`,
+      );
+    }
+    this.#registerActionHandler(actionType, handler);
+    if (this.#parent) {
+      // @ts-expect-error The parent type isn't constructed in a way that proves it supports this
+      // action, but this is OK because it's validated in the constructor.
+      this.delegate({ actions: [actionType], messenger: this.#parent });
+    }
+  }
+
+  #registerActionHandler<ActionType extends Action['type']>(
     actionType: ActionType,
-    handler: ActionHandler<Action, ActionType>,
+    handler: ActionHandler<ActionConstraint, ActionType>,
   ) {
     if (this.#actions.has(actionType)) {
       throw new Error(
@@ -198,8 +316,8 @@ export class Messenger<
    * @template MethodNames - The type union of method names to register as action handlers.
    */
   registerMethodActionHandlers<
-    MessengerClient extends { name: string },
-    MethodNames extends keyof MessengerClient & string,
+    MessengerClient extends { name: Namespace },
+    MethodNames extends keyof MessengerClient & StripNamespace<Action['type']>,
   >(messengerClient: MessengerClient, methodNames: readonly MethodNames[]) {
     for (const methodName of methodNames) {
       const method = messengerClient[methodName];
@@ -215,13 +333,37 @@ export class Messenger<
    *
    * This will prevent this action from being called.
    *
+   * The action being unregistered must be under the same namespace as the messenger.
+   *
    * @param actionType - The action type. This is a unique identifier for this action.
-   * @template ActionType - A type union of Action type strings.
+   * @template ActionType - A type union of Action type strings under this messenger's namespace.
    */
-  unregisterActionHandler<ActionType extends Action['type']>(
+  unregisterActionHandler<
+    ActionType extends Action['type'] & NamespacedName<Namespace>,
+  >(actionType: ActionType) {
+    /* istanbul ignore if */ // Branch unreachable with valid types
+    if (!this.#isInCurrentNamespace(actionType)) {
+      throw new Error(
+        `Only allowed unregistering action handlers prefixed by '${
+          this.#namespace
+        }:'`,
+      );
+    }
+    this.#unregisterActionHandler(actionType);
+  }
+
+  #unregisterActionHandler<ActionType extends Action['type']>(
     actionType: ActionType,
   ) {
     this.#actions.delete(actionType);
+    const delegationTargets = this.#actionDelegationTargets.get(actionType);
+    if (!delegationTargets) {
+      return;
+    }
+    for (const messenger of delegationTargets) {
+      messenger._internalUnregisterDelegatedActionHandler(actionType);
+    }
+    this.#actionDelegationTargets.delete(actionType);
   }
 
   /**
@@ -230,7 +372,9 @@ export class Messenger<
    * This prevents all actions from being called.
    */
   clearActions() {
-    this.#actions.clear();
+    for (const actionType of this.#actions.keys()) {
+      this.#unregisterActionHandler(actionType);
+    }
   }
 
   /**
@@ -267,12 +411,42 @@ export class Messenger<
    * Registering a function for getting the payload allows event selectors to have a point of
    * comparison the first time state changes.
    *
+   * The event type must be under the same namespace as the messenger.
+   *
    * @param args - The arguments to this function
    * @param args.eventType - The event type to register a payload for.
    * @param args.getPayload - A function for retrieving the event payload.
-   * @template EventType - A type union of Event type strings.
+   * @template EventType - A type union of Event type strings under this messenger's namespace.
    */
-  registerInitialEventPayload<EventType extends Event['type']>({
+  registerInitialEventPayload<
+    EventType extends Event['type'] & NamespacedName<Namespace>,
+  >({
+    eventType,
+    getPayload,
+  }: {
+    eventType: EventType;
+    getPayload: () => ExtractEventPayload<Event, EventType>;
+  }) {
+    /* istanbul ignore if */ // Branch unreachable with valid types
+    if (!this.#isInCurrentNamespace(eventType)) {
+      throw new Error(
+        `Only allowed registering initial payloads for events prefixed by '${
+          this.#namespace
+        }:'`,
+      );
+    }
+    if (
+      this.#parent &&
+      !this.#subscriptionDelegationTargets.get(eventType)?.has(this.#parent)
+    ) {
+      // @ts-expect-error The parent type isn't constructed in a way that proves it supports this
+      // event, but this is OK because it's validated in the constructor.
+      this.delegate({ events: [eventType], messenger: this.#parent });
+    }
+    this.#registerInitialEventPayload({ eventType, getPayload });
+  }
+
+  #registerInitialEventPayload<EventType extends Event['type']>({
     eventType,
     getPayload,
   }: {
@@ -280,6 +454,17 @@ export class Messenger<
     getPayload: () => ExtractEventPayload<Event, EventType>;
   }) {
     this.#initialEventPayloadGetters.set(eventType, getPayload);
+    const delegationTargets =
+      this.#subscriptionDelegationTargets.get(eventType);
+    if (!delegationTargets) {
+      return;
+    }
+    for (const messenger of delegationTargets.keys()) {
+      messenger._internalRegisterDelegatedInitialEventPayload({
+        eventType,
+        getPayload,
+      });
+    }
   }
 
   /**
@@ -290,12 +475,35 @@ export class Messenger<
    * Note that this method should never throw directly. Any errors from
    * subscribers are captured and re-thrown in a timeout handler.
    *
+   * The event being published must be under the same namespace as the messenger.
+   *
    * @param eventType - The event type. This is a unique identifier for this event.
    * @param payload - The event payload. The type of the parameters for each event handler must
    * match the type of this payload.
-   * @template EventType - A type union of Event type strings.
+   * @template EventType - A type union of Event type strings under this messenger's namespace.
    */
-  publish<EventType extends Event['type']>(
+  publish<EventType extends Event['type'] & NamespacedName<Namespace>>(
+    eventType: EventType & NamespacedName<Namespace>,
+    ...payload: ExtractEventPayload<Event, EventType>
+  ) {
+    /* istanbul ignore if */ // Branch unreachable with valid types
+    if (!this.#isInCurrentNamespace(eventType)) {
+      throw new Error(
+        `Only allowed publishing events prefixed by '${this.#namespace}:'`,
+      );
+    }
+    if (
+      this.#parent &&
+      !this.#subscriptionDelegationTargets.get(eventType)?.has(this.#parent)
+    ) {
+      // @ts-expect-error The parent type isn't constructed in a way that proves it supports this
+      // event, but this is OK because it's validated in the constructor.
+      this.delegate({ events: [eventType], messenger: this.#parent });
+    }
+    this.#publish(eventType, ...payload);
+  }
+
+  #publish<EventType extends Event['type']>(
     eventType: EventType,
     ...payload: ExtractEventPayload<Event, EventType>
   ) {
@@ -371,16 +579,10 @@ export class Messenger<
       | SelectorEventHandler<SelectorReturnValue>,
     selector?: SelectorFunction<Event, EventType, SelectorReturnValue>,
   ): void {
-    let subscribers = this.#events.get(eventType);
-    if (!subscribers) {
-      subscribers = new Map();
-      this.#events.set(eventType, subscribers);
-    }
-
     // Widen type of event handler by dropping ReturnType parameter.
     //
     // We need to drop it here because it's used as the parameter to the event handler, and
-    // functions in general are contravarient over the parameter type. This means the type is no
+    // functions in general are contravariant over the parameter type. This means the type is no
     // longer valid once it's added to a broader type union with other handlers (because as far
     // as TypeScript knows, we might call the handler with output from a different selector).
     //
@@ -391,7 +593,7 @@ export class Messenger<
     const widenedHandler = handler as
       | ExtractEventHandler<Event, EventType>
       | SelectorEventHandler;
-    subscribers.set(widenedHandler, { selector });
+    this.#subscribe(eventType, widenedHandler, { delegation: false, selector });
 
     if (selector) {
       const getPayload = this.#initialEventPayloadGetters.get(eventType);
@@ -400,6 +602,31 @@ export class Messenger<
         this.#eventPayloadCache.set(widenedHandler, initialValue);
       }
     }
+  }
+
+  /**
+   * Subscribe to an event.
+   *
+   * @param eventType - The event type. This is a unique identifier for this event.
+   * @param handler - The event handler. The type of the parameters for this event handler must
+   * match the type of the payload for this event type.
+   * @param metadata - Event metadata.
+   * @template SubscribedEvent - The event being subscribed to.
+   * @template SelectorReturnValue - The selector return value.
+   */
+  #subscribe<SubscribedEvent extends EventConstraint>(
+    eventType: SubscribedEvent['type'],
+    handler:
+      | ExtractEventHandler<SubscribedEvent, SubscribedEvent['type']>
+      | SelectorEventHandler,
+    metadata: SubscriptionMetadata<SubscribedEvent>,
+  ): void {
+    let subscribers = this.#events.get(eventType);
+    if (!subscribers) {
+      subscribers = new Map();
+      this.#events.set(eventType, subscribers);
+    }
+    subscribers.set(handler, metadata);
   }
 
   /**
@@ -424,7 +651,7 @@ export class Messenger<
     // Widen type of event handler by dropping ReturnType parameter.
     //
     // We need to drop it here because it's used as the parameter to the event handler, and
-    // functions in general are contravarient over the parameter type. This means the type is no
+    // functions in general are contravariant over the parameter type. This means the type is no
     // longer valid once it's added to a broader type union with other handlers (because as far
     // as TypeScript knows, we might call the handler with output from a different selector).
     //
@@ -432,12 +659,14 @@ export class Messenger<
     const widenedHandler = handler as
       | ExtractEventHandler<Event, EventType>
       | SelectorEventHandler;
-    if (!subscribers || !subscribers.has(widenedHandler)) {
+    if (!subscribers) {
       throw new Error(`Subscription not found for event: ${eventType}`);
     }
-
     const metadata = subscribers.get(widenedHandler);
-    if (metadata?.selector) {
+    if (!metadata) {
+      throw new Error(`Subscription not found for event: ${eventType}`);
+    }
+    if (metadata.selector) {
       this.#eventPayloadCache.delete(widenedHandler);
     }
 
@@ -447,7 +676,8 @@ export class Messenger<
   /**
    * Clear subscriptions for a specific event.
    *
-   * This will remove all subscribed handlers for this event.
+   * This will remove all subscribed handlers for this event registered from this messenger. The
+   * event may still have subscribers if it has been delegated to another messenger.
    *
    * @param eventType - The event type. This is a unique identifier for this event.
    * @template EventType - A type union of Event type strings.
@@ -455,75 +685,318 @@ export class Messenger<
   clearEventSubscriptions<EventType extends Event['type']>(
     eventType: EventType,
   ) {
-    this.#events.delete(eventType);
+    const subscriptions = this.#events.get(eventType);
+    if (!subscriptions) {
+      return;
+    }
+
+    for (const [handler, metadata] of subscriptions.entries()) {
+      if (metadata.delegation) {
+        continue;
+      }
+      subscriptions.delete(handler);
+    }
+
+    if (subscriptions.size === 0) {
+      this.#events.delete(eventType);
+    }
   }
 
   /**
    * Clear all subscriptions.
    *
-   * This will remove all subscribed handlers for all events.
+   * This will remove all subscribed handlers for all events registered from this messenger. Events
+   * may still have subscribers if they are delegated to another messenger.
    */
   clearSubscriptions() {
-    this.#events.clear();
+    for (const eventType of this.#events.keys()) {
+      this.clearEventSubscriptions(eventType);
+    }
   }
 
   /**
-   * Get a restricted messenger
+   * Delegate actions and/or events to another messenger.
    *
-   * Returns a wrapper around the messenger instance that restricts access to actions and events.
-   * The provided allowlists grant the ability to call the listed actions and subscribe to the
-   * listed events. The "name" provided grants ownership of any actions and events under that
-   * namespace. Ownership allows registering actions and publishing events, as well as
-   * unregistering actions and clearing event subscriptions.
+   * The messenger these actions/events are delegated to will be able to call these actions and
+   * subscribe to these events.
    *
-   * @param options - Messenger options.
-   * @param options.name - The name of the thing this messenger will be handed to (e.g. the
-   * controller name). This grants "ownership" of actions and events under this namespace to the
-   * restricted messenger returned.
-   * @param options.allowedActions - The list of actions that this restricted messenger should be
-   * allowed to call.
-   * @param options.allowedEvents - The list of events that this restricted messenger should be
-   * allowed to subscribe to.
-   * @template Namespace - The namespace for this messenger. Typically this is the name of the
-   * module that this messenger has been created for. The authority to publish events and register
-   * actions under this namespace is granted to this restricted messenger instance.
-   * @template AllowedAction - A type union of the 'type' string for any allowed actions.
-   * This must not include internal actions that are in the messenger's namespace.
-   * @template AllowedEvent - A type union of the 'type' string for any allowed events.
-   * This must not include internal events that are in the messenger's namespace.
-   * @returns The restricted messenger.
+   * Note that the messenger these actions/events are delegated to must still have these
+   * actions/events included in its type definition (as part of the Action and Event type
+   * parameters). Actions and events are statically type checked, they cannot be delegated
+   * dynamically at runtime.
+   *
+   * @param args - Arguments.
+   * @param args.actions - The action types to delegate.
+   * @param args.events - The event types to delegate.
+   * @param args.messenger - The messenger to delegate to.
+   * @template Delegatee - The messenger the actions/events are delegated to.
+   * @template DelegatedActions - An array of delegated action types.
+   * @template DelegatedEvents - An array of delegated event types.
    */
-  getRestricted<
-    Namespace extends string,
-    AllowedAction extends NotNamespacedBy<Namespace, Action['type']> = never,
-    AllowedEvent extends NotNamespacedBy<Namespace, Event['type']> = never,
+  delegate<
+    Delegatee extends Messenger<string, ActionConstraint, EventConstraint>,
+    DelegatedActions extends (MessengerActions<Delegatee> & Action)['type'][],
+    DelegatedEvents extends (MessengerEvents<Delegatee> & Event)['type'][],
   >({
-    name,
-    allowedActions,
-    allowedEvents,
+    actions,
+    events,
+    messenger,
   }: {
-    name: Namespace;
-    allowedActions: NotNamespacedBy<
-      Namespace,
-      Extract<Action['type'], AllowedAction>
-    >[];
-    allowedEvents: NotNamespacedBy<
-      Namespace,
-      Extract<Event['type'], AllowedEvent>
-    >[];
-  }): RestrictedMessenger<
-    Namespace,
-    | NarrowToNamespace<Action, Namespace>
-    | NarrowToAllowed<Action, AllowedAction>,
-    NarrowToNamespace<Event, Namespace> | NarrowToAllowed<Event, AllowedEvent>,
-    AllowedAction,
-    AllowedEvent
-  > {
-    return new RestrictedMessenger({
-      messenger: this,
-      name,
-      allowedActions,
-      allowedEvents,
-    });
+    actions?: DelegatedActions;
+    events?: DelegatedEvents;
+    messenger: Delegatee;
+  }) {
+    for (const actionType of actions || []) {
+      const delegatedActionHandler = (
+        ...args: ExtractActionParameters<
+          MessengerActions<Delegatee> & Action,
+          typeof actionType
+        >
+      ) => {
+        // Cast to get more specific type, for this specific action
+        // The types get collapsed by `this.#actions`
+        const actionHandler = this.#actions.get(actionType) as
+          | ActionHandler<
+              MessengerActions<Delegatee> & Action,
+              typeof actionType
+            >
+          | undefined;
+        if (!actionHandler) {
+          throw new Error(
+            `Cannot call '${actionType}', action not registered.`,
+          );
+        }
+        return actionHandler(...args);
+      };
+      let delegationTargets = this.#actionDelegationTargets.get(actionType);
+      if (!delegationTargets) {
+        delegationTargets = new Set<DelegatedMessenger>();
+        this.#actionDelegationTargets.set(actionType, delegationTargets);
+      }
+      if (delegationTargets.has(messenger)) {
+        throw new Error(
+          `The action '${actionType}' has already been delegated to this messenger`,
+        );
+      }
+      delegationTargets.add(messenger);
+
+      messenger._internalRegisterDelegatedActionHandler(
+        actionType,
+        delegatedActionHandler,
+      );
+    }
+    for (const eventType of events || []) {
+      const untypedSubscriber = (
+        ...payload: ExtractEventPayload<
+          MessengerEvents<Delegatee> & Event,
+          typeof eventType
+        >
+      ) => {
+        messenger._internalPublishDelegated(eventType, ...payload);
+      };
+      // Cast to get more specific subscriber type for this specific event.
+      // The types get collapsed here to the type union of all delegated
+      // events, rather than the single subscriber type corresponding to this
+      // event.
+      const subscriber = untypedSubscriber as ExtractEventHandler<
+        MessengerEvents<Delegatee> & Event,
+        typeof eventType
+      >;
+      let delegatedEventSubscriptions =
+        this.#subscriptionDelegationTargets.get(eventType);
+      if (!delegatedEventSubscriptions) {
+        delegatedEventSubscriptions = new Map();
+        this.#subscriptionDelegationTargets.set(
+          eventType,
+          delegatedEventSubscriptions,
+        );
+      }
+      if (delegatedEventSubscriptions.has(messenger)) {
+        throw new Error(
+          `The event '${eventType}' has already been delegated to this messenger`,
+        );
+      }
+      delegatedEventSubscriptions.set(messenger, subscriber);
+      const getPayload = this.#initialEventPayloadGetters.get(eventType);
+      if (getPayload) {
+        messenger._internalRegisterDelegatedInitialEventPayload({
+          eventType,
+          getPayload,
+        });
+      }
+
+      this.#subscribe(eventType, subscriber, { delegation: true });
+    }
+  }
+
+  /**
+   * Revoke delegated actions and/or events from another messenger.
+   *
+   * The messenger these actions/events are delegated to will no longer be able to call these
+   * actions or subscribe to these events.
+   *
+   * @param args - Arguments.
+   * @param args.actions - The action types to revoke.
+   * @param args.events - The event types to revoke.
+   * @param args.messenger - The messenger these actions/events were delegated to.
+   * @template Delegatee - The messenger the actions/events are being revoked from.
+   * @template DelegatedActions - An array of delegated action types.
+   * @template DelegatedEvents - An array of delegated event types.
+   */
+  revoke<
+    Delegatee extends Messenger<string, ActionConstraint, EventConstraint>,
+    DelegatedActions extends (MessengerActions<Delegatee> & Action)['type'][],
+    DelegatedEvents extends (MessengerEvents<Delegatee> & Event)['type'][],
+  >({
+    actions,
+    events,
+    messenger,
+  }: {
+    actions?: DelegatedActions;
+    events?: DelegatedEvents;
+    messenger: Delegatee;
+  }) {
+    if (messenger === this.#parent) {
+      throw new Error('Cannot revoke from parent');
+    }
+    for (const actionType of actions || []) {
+      const delegationTargets = this.#actionDelegationTargets.get(actionType);
+      if (!delegationTargets || !delegationTargets.has(messenger)) {
+        // Nothing to revoke
+        continue;
+      }
+      messenger._internalUnregisterDelegatedActionHandler(actionType);
+      delegationTargets.delete(messenger);
+      if (delegationTargets.size === 0) {
+        this.#actionDelegationTargets.delete(actionType);
+      }
+    }
+    for (const eventType of events || []) {
+      const delegationTargets =
+        this.#subscriptionDelegationTargets.get(eventType);
+      if (!delegationTargets) {
+        // Nothing to revoke
+        continue;
+      }
+      const delegatedSubscriber = delegationTargets.get(messenger);
+      if (!delegatedSubscriber) {
+        // Nothing to revoke
+        continue;
+      }
+      this.unsubscribe(eventType, delegatedSubscriber);
+      delegationTargets.delete(messenger);
+      if (delegationTargets.size === 0) {
+        this.#subscriptionDelegationTargets.delete(eventType);
+      }
+    }
+  }
+
+  /**
+   * Register an action handler for an action delegated from another messenger.
+   *
+   * This will make the registered function available to call via the `call` method.
+   *
+   * Note: This is an internal method. Never access this property from another module. This must be
+   * exposed as a public property so that these methods can be called internally on other messenger
+   * instances.
+   *
+   * @deprecated Internal use only. Use the `delegate` method for delegation.
+   * @param actionType - The action type. This is a unique identifier for this action.
+   * @param handler - The action handler. This function gets called when the `call` method is
+   * invoked with the given action type.
+   * @throws Will throw when a handler has been registered for this action type already.
+   * @template ActionType - A type union of Action type strings.
+   */
+  _internalRegisterDelegatedActionHandler<ActionType extends Action['type']>(
+    actionType: ActionType,
+    // Using wider `ActionConstraint` type here rather than `Action` because the `Action` type is
+    // contravariant over the handler parameter type. Using `Action` would lead to a type error
+    // here because the messenger we've delegated to supports _additional_ actions.
+    handler: ActionHandler<ActionConstraint, ActionType>,
+  ) {
+    this.#registerActionHandler(actionType, handler);
+  }
+
+  /**
+   * Unregister an action handler for an action delegated from another messenger.
+   *
+   * This will prevent this action from being called.
+   *
+   * Note: This is an internal method. Never access this property from another module. This must be
+   * exposed as a public property so that these methods can be called internally on other messenger
+   * instances.
+   *
+   * @deprecated Internal use only. Use the `delegate` method for delegation.
+   * @param actionType - The action type. This is a unqiue identifier for this action.
+   * @template ActionType - A type union of Action type strings.
+   */
+  _internalUnregisterDelegatedActionHandler<ActionType extends Action['type']>(
+    actionType: ActionType,
+  ) {
+    this.#unregisterActionHandler(actionType);
+  }
+
+  /**
+   * Register a function for getting the initial payload for an event that has been delegated from
+   * another messenger.
+   *
+   * This is used for events that represent a state change, where the payload is the state.
+   * Registering a function for getting the payload allows event selectors to have a point of
+   * comparison the first time state changes.
+   *
+   * Note: This is an internal method. Never access this property from another module. This must be
+   * exposed as a public property so that these methods can be called internally on other messenger
+   * instances.
+   *
+   * @deprecated Internal use only. Use the `delegate` method for delegation.
+   * @param args - The arguments to this function
+   * @param args.eventType - The event type to register a payload for.
+   * @param args.getPayload - A function for retrieving the event payload.
+   */
+  _internalRegisterDelegatedInitialEventPayload<
+    EventType extends Event['type'],
+  >({
+    eventType,
+    getPayload,
+  }: {
+    eventType: EventType;
+    getPayload: () => ExtractEventPayload<Event, EventType>;
+  }) {
+    this.#registerInitialEventPayload({ eventType, getPayload });
+  }
+
+  /**
+   * Publish an event that was delegated from another messenger.
+   *
+   * Publishes the given payload to all subscribers of the given event type.
+   *
+   * Note that this method should never throw directly. Any errors from
+   * subscribers are captured and re-thrown in a timeout handler.
+   *
+   * Note: This is an internal method. Never access this property from another module. This must be
+   * exposed as a public property so that these methods can be called internally on other messenger
+   * instances.
+   *
+   * @deprecated Internal use only. Use the `delegate` method for delegation.
+   * @param eventType - The event type. This is a unique identifier for this event.
+   * @param payload - The event payload. The type of the parameters for each event handler must
+   * match the type of this payload.
+   * @template EventType - A type union of Event type strings.
+   */
+  _internalPublishDelegated<EventType extends Event['type']>(
+    eventType: EventType,
+    ...payload: ExtractEventPayload<Event, EventType>
+  ) {
+    this.#publish(eventType, ...payload);
+  }
+
+  /**
+   * Determine whether the given name is within the current namespace.
+   *
+   * @param name - The name to check
+   * @returns Whether the name is within the current namespace
+   */
+  #isInCurrentNamespace(name: string): name is NamespacedName<Namespace> {
+    return name.startsWith(`${this.#namespace}:`);
   }
 }